PostgreSQL Database Integration (Optional)
I've added a PostgreSQL database to manage food items and user login details (emails, OTPs, passwords). It's not required to run the project since you can see the food items directly on home_page.html, but if you're curious, this is where the data comes from and gets stored when the admin submits new info.